+++ API/ParserInterface.php
@@ -55,7 +55,7 @@ interface ParserInterface
      *
      * @return string Processed styles, in inline string
      */
-    public function inlineStylesData(array $styles, string $route, string $id);
+    public function processStylesData(array $styles, string $route, string $id);
 
     /**
      * Remove wrapping paragraph from img-element